requirements.txt
requests
aiohttp==3.9.1
python-dotenv==1.0.0
supabase==2.3.0
asyncio==3.4.3